+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to `make-action` will be documented in this file.
 
 ## 1.0.0 - 2025-07-21
 
+- Added `make:action` command.
 - Updated package namespace from `GlenBangkila\MakeAction` to `Webteractive\MakeAction`.
 - Added Laravel Prompts support for interactive command input.
 - Added `laravel/prompts` as a dev dependency.